+++ src/Informedica.GenORDER.Lib/Order.fs
@@ -14,7 +14,6 @@ module Order =
     open System
     open Informedica.Utils.Lib.BCL
     open Informedica.Utils.Lib
-    open ConsoleWriter.NewLineNoTime
     open Informedica.GenUnits.Lib
     open WrappedString
 
@@ -1887,10 +1886,11 @@ module Order =
         /// Increase the Quantity increment of an Orderable to a maximum
         /// count using a list of increments.
         /// </summary>
+        /// <param name="logger">The logger</param>
         /// <param name="maxCount">The maximum count</param>
         /// <param name="incrs">The list of increments</param>
         /// <param name="orb">The Orderable</param>
-        let increaseQuantityIncrement maxCount incrs orb =
+        let increaseQuantityIncrement logger maxCount incrs orb =
             // check if all relevant OrderVariables have an increment
             if
                 (orb |> inf).Components
@@ -1919,8 +1919,12 @@ module Order =
                                     incr |> Variable.ValueRange.Increment.toValueUnit |> ValueUnit.getBaseValue
                                 )
 
-                            writeDebugMessage
-                                $"Increase quantity increment to {incr |> Variable.ValueRange.Increment.toString false}"
+                            Logging.logDebugLazy
+                                logger
+                                (fun () ->
+                                    $"Increase quantity increment to {incr |> Variable.ValueRange.Increment.toString false}"
+                                    |> Events.OrderScenario
+                                )
                             // apply the minimum increment increase to the orderable and components
                             { orb with
                                 OrderableQuantity =
@@ -2818,8 +2822,10 @@ module Order =
         |> List.filter String.notEmpty
 
 
-    let print ord =
-        ord |> toStringWithConstraints |> String.concat "\n" |> writeDebugMessage
+    let print logger ord =
+        Logging.logDebugLazy
+            logger
+            (fun () -> ord |> toStringWithConstraints |> String.concat "\n" |> Events.OrderScenario)
 
         ord
 
@@ -3164,11 +3170,12 @@ module Order =
     /// Increase the Quantity increment of an Order to a maximum
     /// count using a list of increments.
     /// </summary>
+    /// <param name="logger">The logger</param>
     /// <param name="maxCount">The maximum count</param>
     /// <param name="incrs">The list of increments</param>
     /// <param name="ord">The Order</param>
-    let increaseQuantityIncrement maxCount incrs ord =
-        { (ord |> inf) with Orderable = ord.Orderable |> Orderable.increaseQuantityIncrement maxCount incrs }
+    let increaseQuantityIncrement logger maxCount incrs ord =
+        { (ord |> inf) with Orderable = ord.Orderable |> Orderable.increaseQuantityIncrement logger maxCount incrs }
 
 
     /// <summary>
@@ -3408,19 +3415,28 @@ module Order =
                     |> mapFromOrderEquations ord
                     |> fun ord ->
                         if printErr then
-                            writeDebugMessage $"Solve errored with: {m}"
-                            ord |> toString |> List.mapi (sprintf "%i. %s") |> List.iter writeDebugMessage
+                            Logging.logDebugLazy
+                                logger
+                                (fun () ->
+                                    $"""Solve errored with: {m}
+{ord |> toString |> List.mapi (sprintf "%i. %s") |> String.concat "\n"}"""
+                                    |> Events.OrderScenario
+                                )
 
                         Error(ord, m)
         with exn ->
             if printErr then
-                writeDebugMessage $"Solve errored with: {exn.Message}"
-
-                oEqs
-                |> mapFromOrderEquations ord
-                |> toString
-                |> List.mapi (sprintf "%i. %s")
-                |> List.iter writeDebugMessage
+                Logging.logDebugLazy
+                    logger
+                    (fun () ->
+                        $"""Solve errored with: {exn.Message}
+{oEqs
+ |> mapFromOrderEquations ord
+ |> toString
+ |> List.mapi (sprintf "%i. %s")
+ |> String.concat "\n"}"""
+                        |> Events.OrderScenario
+                    )
 
             let msg = [ exn |> Informedica.GenSolver.Lib.Types.Exceptions.UnexpectedException ]
 
@@ -3485,7 +3501,9 @@ module Order =
             then
                 ord
             else
-                let incrOrd = ord |> increaseQuantityIncrement maxQtyCount (incrs Units.Volume.milliLiter)
+                let incrOrd =
+                    ord
+                    |> increaseQuantityIncrement logger maxQtyCount (incrs Units.Volume.milliLiter)
 
                 if ord = incrOrd then
                     Ok ord
@@ -3495,8 +3513,14 @@ module Order =
                     incrOrd |> solveMinMax "Increase Quantity Increment" false logger
                 |> function
                     | Error(_, errs) ->
-                        writeDebugMessage "Could not increase orderable quantity increment:"
-                        errs |> List.iter (fun e -> writeDebugMessage $"{e}")
+                        Logging.logDebugLazy
+                            logger
+                            (fun () ->
+                                $"""Could not increase orderable quantity increment:
+{errs |> List.map string |> String.concat "\n"}"""
+                                |> Events.OrderScenario
+                            )
+
                         ord // original order
                     | Ok ord ->
                         let incrOrd =
@@ -3514,8 +3538,14 @@ module Order =
 
                         |> function
                             | Error(_, errs) ->
-                                writeDebugMessage "Could not increase orderable rate increment:"
-                                errs |> List.iter (fun e -> writeDebugMessage $"{e}")
+                                Logging.logDebugLazy
+                                    logger
+                                    (fun () ->
+                                        $"""Could not increase orderable rate increment:
+{errs |> List.map string |> String.concat "\n"}"""
+                                        |> Events.OrderScenario
+                                    )
+
                                 ord // increased increment order
                             | Ok ord -> ord // increased increment and rate order
         |> Ok
@@ -3545,7 +3575,13 @@ module Order =
                 |> fromOrdVars ovars
                 |> solveOrder "Maximize Rate" false logger
                 |> Result.map (fun ord ->
-                    writeDebugMessage $"max rate set to: {maxRte |> OrderVariable.toString true}"
+                    Logging.logDebugLazy
+                        logger
+                        (fun () ->
+                            $"max rate set to: {maxRte |> OrderVariable.toString true}"
+                            |> Events.OrderScenario
+                        )
+
                     ord
                 )
                 |> Result.defaultValue ord
@@ -3643,7 +3679,9 @@ module Order =
                         isSolved <- true
                         loop ord
                     | Error err ->
-                        err |> snd |> List.map (sprintf "%A") |> String.concat "\n" |> writeErrorMessage
+                        (err |> snd |> List.map (sprintf "%A") |> String.concat "\n", err |> fst)
+                        |> Exceptions.OrderCouldNotBeSolved
+                        |> Logging.logError logger
 
                         ord
 
@@ -4439,8 +4477,6 @@ module Order =
 
                 create id adj_qty orb sch dto.Route ord_tme sts |> Ok
             with exn ->
-                $"Could not create an Order from a dto with:\n{exn}" |> writeErrorMessage
-
                 exn |> Exceptions.OrderCouldNotBeCreated |> Error